requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.104.0
python-dateutil==2.8.2
